test(core): split columnar_index's tests into a sibling file (module-size ratchet)
Maintainer (louistrue) on PR #2882: this PR's new tests pushed
rust/core/src/columnar_index.rs to 436 lines, over the module_size_ratchet's
400-line budget, with no allowlist row — failing the required `Rust tests`
CI lane (`no_module_grows_past_its_ratchet_budget`). Per the house rule the
maintainer cited, prefer splitting over raising the budget.

Move the `#[cfg(test)] mod columnar_index_tests { ... }` block out to
rust/core/src/columnar_index_tests.rs, wired back in via
`#[path = "columnar_index_tests.rs"]`, mirroring the existing
georef.rs/georef_tests.rs split in this same crate. columnar_index.rs is now
317 lines; no allowlist change needed.

Verified: `cargo test -p ifc-lite-processing --test module_size_ratchet`
passes; the same 7 columnar_index unit tests still pass unchanged
(cargo test -p ifc-lite-core --lib columnar_index); full
`cargo test --workspace --no-fail-fast` is green; `cargo clippy --workspace
--exclude ifc-lite-wasm --all-targets -- -D warnings` is clean — the exact
commands the maintainer's comment asked to run before pushing.
